refactor: 공통 컴포넌트 제외한 나머지 컴포넌트 container/contracts/components로 이동
