+++ frontend/src/container/contracts/components/ContractTerms.tsx
